Error View Addition and Dynamic Profile Pages
Added implementation for changes in profile pages where certain elements
will only be shown when a user is logged in or looking at another user's
profile.
Also added a new view model for Errors in which it will display an error
page. This is used for when a search for a username cannot be found or
someone inputs an incorrect profileId in the query param of the URL.
Added search functionality for usernames that will redirect to the page
when found.